fix: extend MoE-aware routing to qwen3_vl and glm4v base models
Add MoE config fields with safe zero/None defaults to qwen3_vl and
glm4v TextConfig so these base models handle MoE weights when present.

For qwen3_vl: add SparseMoeBlock, make DecoderLayer check num_experts > 0
to select MoE vs dense MLP. Simplify qwen3_vl_moe to thin re-exports.

For glm4v: add MoEGate, MoE, group_expert_select. Make DecoderLayer check
n_routed_experts to select MoE vs dense MLP. Simplify glm4v_moe to thin
re-exports.

Same pattern as the qwen3_5 fix in the previous commit.
